Changes in listView classes. Fixed a few minor bugs.
